4th branch commit
dtos:
- Created UserDTO
- Created AttendantDTO, extends from UserDTO

DTOs (Data Transfer Objects) are used to get and set data instead of the
entities themselves, this way no data is stored in the entities instead
it goes to the DTOs

entities:
- Changed the id type from int to long

Web Pages:
- Created admin_attendants_create.xhtml, it allows the creation of new
attendants
- Modified index.xhtml to show all exisiting attendants, as well as a
link to the attendants creation page

ejbs:
- Modified the ConfigBean to populate the database with 4 attendants for
display on the index.xhtml

web:
- Modified the AdministratorManager in order to set-up the attendant
creation page, it now receives the input from the webpage and creates an
attendant with that data

+++ AttendanceManagement/src/java/entities/Administrator.java
@@ -20,7 +20,7 @@ public class Administrator extends User implements Serializable {
     public Administrator() {
     }
 
-    public Administrator(int id, String name, String password, String email) {
+    public Administrator(long id, String name, String password, String email) {
         super(id, name, password, email);
     }
